Improved reset/load abilities of lineItem Classes
Added resetItemData()  and updated loadItemData() in the lineItem Class. This was done to fix a bug with the receipt data persisting between different receipt reviews. In addition the new resetItemData has been overridden in the Allocations class so that allocationdata can be reset to a default from the menu system.